update SwiftUI buttons
